+++ src/main/java/se/java/security/models/Availability.java
@@ -1,21 +1,15 @@
 package se.java.security.models;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import org.springframework.data.annotation.Id;
-
 import java.time.LocalDate;
-import java.util.Date;
+
 
 public class Availability {
     private Booking booking;
     @NotNull(message = "Start date must not be null")
-    @JsonFormat(pattern = "yyyy-MM-dd")
     private LocalDate startDate;
 
     @NotNull(message = "End date must not be null")
-    @JsonFormat(pattern = "yyyy-MM-dd")
     private LocalDate endDate;
 
     public Availability(LocalDate startDate, LocalDate endDate) {
