+++ src/main/java/se/java/security/models/Rating.java
@@ -1,7 +1,9 @@
 package se.java.security.models;
 
 import jakarta.validation.constraints.*;
+import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.time.LocalDateTime;
@@ -9,57 +11,55 @@ import java.time.LocalDateTime;
 @Document(collection = "ratings")
 public class Rating {
     @Id
-    private String ratingId;
+    private String id;
 
     @NotBlank(message = "userId cannot be empty")
-    private String UserId;
+
+    private User userId;
 
     @NotBlank(message = "listingId cannot be empty")
-    private String listingId;
+    @DBRef
+    private Listing listingId;
 
     @Min(value = 1, message = "Rating must be at least 1")
     @Max(value = 5, message = "Rating must be at most 5")
     private int rating;
 
+    @CreatedDate
     private LocalDateTime createdAt;
 
-    private LocalDateTime updatedAt;
-
-    public Rating(String ratingId, String UserId, String listingId, int rating, LocalDateTime createdAt, LocalDateTime updatedAt) {
-        this.ratingId = ratingId;
-        this.UserId = UserId;
+    public Rating(String id, User userId, Listing listingId, int rating, LocalDateTime createdAt) {
+        this.id = id;
+        this.userId = userId;
         this.listingId = listingId;
         this.rating = rating;
         this.createdAt = createdAt;
-        this.updatedAt = updatedAt;
     }
 
     public Rating() {
-        this.createdAt = LocalDateTime.now();
-        this.updatedAt = LocalDateTime.now();
     }
 
     public String getId() {
-        return ratingId;
+        return id;
     }
 
-    public void setId(String ratingId) {
-        this.ratingId = ratingId;
+    public void setId(String id) {
+        this.id = id;
     }
 
-    public String getUserId() {
-        return UserId;
+    public User getUserId() {
+        return userId;
     }
 
-    public void setUserId(String userId) {
-        this.UserId = userId;
+    public void setUserId(User userId) {
+        this.userId = userId;
     }
 
-    public String getListingId() {
+    public Listing getListingId() {
         return listingId;
     }
 
-    public void setListingId(String listingId) {
+    public void setListingId(Listing listingId) {
         this.listingId = listingId;
     }
 
@@ -79,11 +79,4 @@ public class Rating {
         this.createdAt = createdAt;
     }
 
-    public LocalDateTime getUpdatedAt() {
-        return updatedAt;
-    }
-
-    public void setUpdatedAt(LocalDateTime updatedAt) {
-        this.updatedAt = updatedAt;
-    }
 }
